refactor: :truck: rename boards.rotes.tx to boards.ts
rename to keep consistency in the project
